Refactor to extract web client
