+++ .gitignore
@@ -1,0 +1,2 @@
+config.py
+*.pyc
