+++ config.sample.py
@@ -1,0 +1,8 @@
+# this file is used by the run.py test script
+# it is not required by the custom component
+
+# copy this file to config.py and replace the values
+
+USERNAME = "replace"
+PASSWORD = "replace"
+DEVICE_ID = "replace"
